+++ flask_via/examples/include/bar/routes.py
@@ -8,10 +8,10 @@ A include ``Flask-Via`` example Flask application.
 """
 
 from flask_via.examples.include.bar.views import FooView, FazView, flop
-from flask.ext.via.routers.default import Basic, Pluggable
+from flask.ext.via.routers.default import Functional, Pluggable
 
 routes = [
     Pluggable('/foo', FooView, 'foo'),
     Pluggable('/faz', FazView, 'faz'),
-    Basic('/flop', flop),
+    Functional('/flop', flop),
 ]